+++ mreg_cli/commands/permission.py
@@ -2,30 +2,43 @@
 
 import argparse
 import ipaddress
+from typing import Any
 
-from .cli import Flag, cli
-from .history import history
-from .log import cli_info, cli_warning
-from .outputmanager import OutputManager
-from .types import IP_network
-from .util import convert_wildcard_to_regex, delete, get, get_list, is_valid_network, patch, post
+from mreg_cli.commands.base import BaseCommand
+from mreg_cli.commands.registry import CommandRegistry
+from mreg_cli.log import cli_info, cli_warning
+from mreg_cli.outputmanager import OutputManager
+from mreg_cli.types import Flag, IP_network
+from mreg_cli.utilities.api import delete, get, get_list, patch, post
+from mreg_cli.utilities.shared import convert_wildcard_to_regex
+from mreg_cli.utilities.validators import is_valid_network
 
-###################################
-#  Add the main command 'access'  #
-###################################
+command_registry = CommandRegistry()
 
-permission = cli.add_command(
-    prog="permission",
-    description="Manage permissions.",
-    short_desc="Manage permissions",
+
+class PermissionCommands(BaseCommand):
+    """Permission commands for the CLI."""
+
+    def __init__(self, cli: Any) -> None:
+        """Initialize the permission commands."""
+        super().__init__(
+            cli, command_registry, "permission", "Manage permission.", "Manage permission"
+        )
+
+
+@command_registry.register_command(
+    prog="network_list",
+    description="List permissions for networks",
+    short_desc="List permissions for networks",
+    flags=[
+        Flag(
+            "-group",
+            description="Group with access (supports wildcards)",
+            metavar="GROUP",
+        ),
+        Flag("-range", description="Network range", metavar="RANGE"),
+    ],
 )
-
-
-##########################################
-# Implementation of sub command 'list' #
-##########################################
-
-
 def network_list(args: argparse.Namespace) -> None:
     """List permissions for networks.
 
@@ -80,26 +93,16 @@ def network_list(args: argparse.Namespace) -> None:
     OutputManager().add_formatted_table(headers, keys, data)
 
 
-permission.add_command(
-    prog="network_list",
-    description="List permissions for networks",
-    short_desc="List permissions for networks",
-    callback=network_list,
+@command_registry.register_command(
+    prog="network_add",
+    description="Add permission for network",
+    short_desc="Add permission for network",
     flags=[
-        Flag(
-            "-group",
-            description="Group with access (supports wildcards)",
-            metavar="GROUP",
-        ),
-        Flag("-range", description="Network range", metavar="RANGE"),
+        Flag("range", description="Network range", metavar="RANGE"),
+        Flag("group", description="Group with access", metavar="GROUP"),
+        Flag("regex", description="Regular expression", metavar="REGEX"),
     ],
 )
-
-##########################################
-# Implementation of sub command 'add' #
-##########################################
-
-
 def network_add(args: argparse.Namespace) -> None:
     """Add permission for network.
 
@@ -114,29 +117,20 @@ def network_add(args: argparse.Namespace) -> None:
         "regex": args.regex,
     }
     path = "/api/v1/permissions/netgroupregex/"
-    history.record_post(path, "", data)
     post(path, **data)
     cli_info(f"Added permission to {args.range}", True)
 
 
-permission.add_command(
-    prog="network_add",
-    description="Add permission for network",
-    short_desc="Add permission for network",
-    callback=network_add,
+@command_registry.register_command(
+    prog="network_remove",
+    description="Remove permission for network",
+    short_desc="Remove permission for network",
     flags=[
         Flag("range", description="Network range", metavar="RANGE"),
         Flag("group", description="Group with access", metavar="GROUP"),
         Flag("regex", description="Regular expression", metavar="REGEX"),
     ],
 )
-
-
-##########################################
-# Implementation of sub command 'remove' #
-##########################################
-
-
 def network_remove(args: argparse.Namespace) -> None:
     """Remove permission for networks.
 
@@ -156,29 +150,21 @@ def network_remove(args: argparse.Namespace) -> None:
     assert len(permissions) == 1, "Should only match one permission"
     identifier = permissions[0]["id"]
     path = f"/api/v1/permissions/netgroupregex/{identifier}"
-    history.record_delete(path, dict(), undoable=False)
     delete(path)
     cli_info(f"Removed permission for {args.range}", True)
 
 
-permission.add_command(
-    prog="network_remove",
-    description="Remove permission for network",
-    short_desc="Remove permission for network",
-    callback=network_remove,
+@command_registry.register_command(
+    prog="label_add",
+    description="Add a label to a permission",
+    short_desc="Add label",
     flags=[
         Flag("range", description="Network range", metavar="RANGE"),
         Flag("group", description="Group with access", metavar="GROUP"),
         Flag("regex", description="Regular expression", metavar="REGEX"),
+        Flag("label", description="The label you want to add"),
     ],
 )
-
-
-#################################################################
-# Implementation of sub commands 'label_add' and 'label_remove'
-#################################################################
-
-
 def add_label_to_permission(args: argparse.Namespace) -> None:
     """Add a label to a permission triplet.
 
@@ -219,19 +205,17 @@ def add_label_to_permission(args: argparse.Namespace) -> None:
     cli_info(f"Added the label {args.label!r} to the permission.", print_msg=True)
 
 
-permission.add_command(
-    prog="label_add",
-    description="Add a label to a permission",
-    callback=add_label_to_permission,
+@command_registry.register_command(
+    prog="label_remove",
+    description="Remove a label from a permission",
+    short_desc="Remove label",
     flags=[
         Flag("range", description="Network range", metavar="RANGE"),
         Flag("group", description="Group with access", metavar="GROUP"),
         Flag("regex", description="Regular expression", metavar="REGEX"),
-        Flag("label", description="The label you want to add"),
+        Flag("label", description="The label you want to remove"),
     ],
 )
-
-
 def remove_label_from_permission(args: argparse.Namespace) -> None:
     """Remove a label from a permission.
 
@@ -270,16 +254,3 @@ def remove_label_from_permission(args: argparse.Namespace) -> None:
     ar.remove(label["id"])
     patch(path, params={"labels": ar}, use_json=True)
     cli_info(f"Removed the label {args.label!r} from the permission.", print_msg=True)
-
-
-permission.add_command(
-    prog="label_remove",
-    description="Remove a label from a permission",
-    callback=remove_label_from_permission,
-    flags=[
-        Flag("range", description="Network range", metavar="RANGE"),
-        Flag("group", description="Group with access", metavar="GROUP"),
-        Flag("regex", description="Regular expression", metavar="REGEX"),
-        Flag("label", description="The label you want to remove"),
-    ],
-)
