+++ mreg_cli/config.py
@@ -14,10 +14,11 @@ translated according to a mapping:
     3. :py:const:`logging.DEBUG`
 """
 
+import configparser
 import logging
 import os
 import sys
-from typing import Tuple, Union
+from typing import Any, Dict, Optional, Tuple, Union
 
 logger = logging.getLogger(__name__)
 
@@ -33,9 +34,6 @@ DEFAULT_CONFIG_PATH = tuple(
     )
 )
 
-DEFAULT_URL = None
-DEFAULT_DOMAIN = None
-
 # Default logging format
 # TODO: Support logging config
 LOGGING_FORMAT = "%(levelname)s - %(name)s - %(message)s"
@@ -49,49 +47,143 @@ LOGGING_VERBOSITY: Tuple[int, int, int, int] = (
 )
 
 
-def get_verbosity(verbosity: int) -> int:
-    """Translate verbosity to logging level.
+class MregCliConfig:
+    """Configuration class for the mreg-cli.
 
-    Levels are traslated according to :py:const:`LOGGING_VERBOSITY`.
-
-    :param int verbosity: verbosity level
-
-    :rtype: int
+    This is a singleton class that is used to store configuration information.
+    Configuration is loaded with the following priority:
+    1. Command line options
+    2. Environment variables (prefixed with 'MREG_')
+    3. Configuration file
     """
-    level = LOGGING_VERBOSITY[min(len(LOGGING_VERBOSITY) - 1, verbosity)]
-    return level
 
+    _instance = None
 
-def configure_logging(level: int = logging.INFO) -> None:
-    """Enable and configure logging.
+    def __new__(cls):
+        """Create a new instance of the configuration class.
 
-    :param int level: logging level, defaults to :py:const:`logging.INFO`
-    """
-    logging.basicConfig(level=level, format=LOGGING_FORMAT)
+        This ensures that only one instance of the configuration class is created.
+        """
+        if cls._instance is None:
+            cls._instance = super().__new__(cls)
+            cls._instance._config_file: Dict[str, str] = {}
+            cls._instance._config_env: Dict[str, str] = cls._load_env_config()
+            cls._instance._config_cmd: Dict[str, str] = {}
+            cls._instance.get_config()
 
+        return cls._instance
 
-def get_config_file() -> Union[str, None]:
-    """Get the first config file found in DEFAULT_CONFIG_PATH.
+    @staticmethod
+    def _load_env_config() -> Dict[str, str]:
+        """Load environment variables into the configuration, filtering with 'MREGCLI_' prefix."""
+        env_prefix = "MREGCLI_"
+        return {
+            key[len(env_prefix) :].lower(): value
+            for key, value in os.environ.items()
+            if key.startswith(env_prefix)
+        }
 
-    :returns: path to config file, or None if no config file was found
-    """
-    for path in DEFAULT_CONFIG_PATH:
-        logger.debug("looking for config in %r", os.path.abspath(path))
-        if os.path.isfile(path):
-            logger.info("found config in %r", path)
-            return path
-    logger.debug("no config file found in config paths")
-    return None
+    def get(self, key: str, default: str = None) -> Optional[str]:
+        """Get a configuration value with priority: cmdline, env, file.
 
+        :param str key: Configuration key.
+        :param default: Default value if key is not found.
 
-def get_default_domain():
-    """Get the default domain from the application."""
-    return DEFAULT_DOMAIN
+        :returns: Configuration value.
+        """
+        return self._config_cmd.get(
+            key, self._config_env.get(key, self._config_file.get(key, default))
+        )
 
+    def set_cmd_config(self, cmd_config: Dict[str, Any]) -> None:
+        """Set command line configuration options.
 
-def get_default_url():
-    """Get the default url from the application."""
-    for url in (os.environ.get("MREGCLI_DEFAULT_URL"), DEFAULT_URL):
-        if url is not None:
-            return url
-    return None
+        :param Dict[str, Any] cmd_config: Dictionary of command line configurations.
+        """
+        self._config_cmd.update(cmd_config)
+
+    def get_config(self, reload: bool = False) -> None:
+        """Load the configuration file into the class.
+
+        :param bool reload: Reload the configuration from the config file.
+        """
+        if not self._config_file or reload:
+            configpath = self.get_config_file()
+            if configpath is not None:
+                cfgparser = configparser.ConfigParser()
+                cfgparser.read(configpath)
+                self._config_file = dict(cfgparser["mreg"].items())
+
+    def get_verbosity(self, verbosity: int) -> int:
+        """Translate verbosity to logging level.
+
+        Levels are traslated according to :py:const:`LOGGING_VERBOSITY`.
+
+        :param int verbosity: verbosity level
+
+        :rtype: int
+        """
+        level = LOGGING_VERBOSITY[min(len(LOGGING_VERBOSITY) - 1, verbosity)]
+        return level
+
+    def configure_logging(self, level: int = logging.INFO) -> None:
+        """Enable and configure logging.
+
+        :param int level: logging level, defaults to :py:const:`logging.INFO`
+        """
+        logging.basicConfig(level=level, format=LOGGING_FORMAT)
+
+    def get_config_file(self) -> Union[str, None]:
+        """Get the first config file found in DEFAULT_CONFIG_PATH.
+
+        :returns: path to config file, or None if no config file was found
+        """
+        for path in DEFAULT_CONFIG_PATH:
+            logger.debug("looking for config in %r", os.path.abspath(path))
+            if os.path.isfile(path):
+                logger.info("found config in %r", path)
+                return path
+        logger.debug("no config file found in config paths")
+        return None
+
+    def get_default_domain(self):
+        """Get the default domain from the application."""
+        return self.get("domain")
+
+    def get_default_url(self):
+        """Get the default url from the application."""
+        self.get("url", self.get("default_url", None))
+
+    def _calculate_column_width(self, data: Dict[str, Any], min_width: int = 8) -> int:
+        """Calculate the maximum column width, ensuring a minimum width.
+
+        :param data: Dictionary of data for the column.
+        :param min_width: Minimum width of the column.
+        :returns: Calculated column width.
+        """
+        max_length = max((len(str(value)) for value in data.values()), default=0)
+        return max(max_length, min_width) + 2  # Adding 2 for padding
+
+    def print_config_table(self) -> None:
+        """Pretty print the configuration options in a dynamic table format."""
+        all_keys = set(self._config_cmd) | set(self._config_env) | set(self._config_file)
+        key_width = max(max(len(key) for key in all_keys), 8) + 2
+
+        # Calculate column widths
+        cmd_width = self._calculate_column_width(self._config_cmd)
+        env_width = self._calculate_column_width(self._config_env)
+        file_width = self._calculate_column_width(self._config_file)
+
+        # Print the table header
+        print("Configuration Options:")
+        header_format = f"{{:<{key_width}}} {{:<{cmd_width}}} {{:<{env_width}}} {{:<{file_width}}}"
+        print(header_format.format("Key", "Active", "Envir", "File"))
+        print("-" * (key_width + cmd_width + env_width + file_width))
+
+        # Print each row
+        row_format = f"{{:<{key_width}}} {{:<{cmd_width}}} {{:<{env_width}}} {{:<{file_width}}}"
+        for key in sorted(all_keys):
+            cmd_line_val = str(self._config_cmd.get(key, "-"))
+            env_var_val = str(self._config_env.get(key, "-"))
+            config_file_val = str(self._config_file.get(key, "-"))
+            print(row_format.format(key, cmd_line_val, env_var_val, config_file_val))
