+++ mreg_cli/utilities/__init__.py
@@ -1,0 +1,5 @@
+"""Utility modules for mreg_cli.
+
+These modules are somewhat logically split based on the scope they typically
+apply to.
+"""
